Disable acceleration of LDIR/LDDR/DJNZ in tap2sna.py
Previously acceleration was enabled only when tracing was disabled,
which could lead to a pause between two tape blocks occurring either
after an LDIR/LDDR/DJNZ instruction (when accelerated), or in the middle
of one (when not accelerated). This inconsistency could in turn lead to
differences in RAM contents or register values in the snapshot produced.